third_party/WebKit/LayoutTests/http/tests/serviceworker/client-focus.html

<!DOCTYPE html>
<title>Service Worker: WindowClient.focus()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var t = async_test('Test WindowClient.focus()');
t.step(function() {
    var scope = 'resources/blank.html'
    service_worker_unregister_and_register(
        t, 'resources/client-focus.js', scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(scope); })
      .then(function(frame) {
          var w = frame.contentWindow;
          w.onmessage = t.step_func(onMessage);
          w.navigator.serviceWorker.controller.postMessage('ping');
        })
      .catch(unreached_rejection(t));

    var result = [];
    var expected = ['focus() is present', 'focus() succeeded with false'];

    function onMessage(e) {
      var message = e.data;
      if (message === 'quit') {
        assert_array_equals(result, expected,
                            'Worker should post back expected messages.');
        service_worker_unregister_and_done(t, scope);
      } else {
        result.push(message);
      }
    }
  });
</script>
